requirements.txt
pyserial
PyYAML
gpsd-py3  # For built-in GPS using GPSD
pytest     # For running test cases
